[Refactor] Enforce per-node write sandbox, add upper bound, and split KB path tests
Finishes addressing CodeRabbit feedback on the knowledge_base_home refactor:

- make_kb_path_normalizer now accepts a strict_kind kwarg. When invoked with
  strict_kind=True (write/edit path), cross-kind writes are rejected
  (e.g. a semantic node may no longer write under sql_summaries/*) and
  file_type overrides are ignored so the node's configured kind always wins.
  Reads stay lax so the LLM can still browse peer KB artifacts.
- FilesystemFuncTool._normalize forwards strict_kind on writes and keeps a
  TypeError fallback for custom normalizers that don't accept the kwarg.
- pyproject.toml: add an upper bound to datus-storage-base (>=0.1.2,<0.2.0)
  so unannounced 0.x breakage cannot slip in.

Tests redistributed per the project's source→test mirror rule:
- tests/unit_tests/cli/test_kb_path_normalization.py is removed.
- Pure normalize_kb_relative_path + make_kb_path_normalizer + hook/tool
  agreement tests moved into tests/unit_tests/cli/test_generation_hooks.py
  (now includes new strict_kind coverage).
- FilesystemFuncTool round-trip tests (including the requested
  read_multiple_files batch case, plus cross-kind write rejection and
  fail-closed normalizer errors) moved into
  tests/unit_tests/tools/func_tool/test_filesystem_tools.py.
- Each of the 4 generation nodes gets a TestX FilesystemRootPath class
  asserting FilesystemFuncTool.config.root_path == knowledge_base_home
  alongside the existing node tests.

Full unit suite: 9256 passed (same 2 pre-existing auth failures on main).
Diff coverage remains above the 80% gate.
