+++ tests/unit_tests/agent/node/test_gen_ext_knowledge_agentic_node.py
@@ -420,3 +420,18 @@ class TestGenExtKnowledgeNodeExecution:
         with pytest.raises(ValueError, match="External knowledge input not set"):
             async for _ in node.execute_stream():
                 pass
+
+
+class TestGenExtKnowledgeFilesystemRootPath:
+    """FilesystemFuncTool is sandboxed to knowledge_base_home (not the type-specific subdir)."""
+
+    def test_filesystem_root_is_kb_home(self, real_agent_config, mock_llm_create):
+        node = _create_node(real_agent_config)
+        expected = str(real_agent_config.path_manager.knowledge_base_home)
+
+        assert node.filesystem_func_tool is not None
+        assert node.filesystem_func_tool.config.root_path == expected
+        assert node.filesystem_func_tool._path_normalizer is not None
+
+        ns = real_agent_config.current_namespace
+        assert node.filesystem_func_tool._path_normalizer("notes.yaml", None) == f"ext_knowledge/{ns}/notes.yaml"
