+++ tests/unit_tests/agent/node/test_sql_summary_agentic_node.py
@@ -318,3 +318,18 @@ class TestSqlSummaryExtractMethods:
         file_name, output = node._extract_sql_summary_and_output_from_response({"content": ""})
         assert file_name is None
         assert output is None
+
+
+class TestSqlSummaryFilesystemRootPath:
+    """FilesystemFuncTool is sandboxed to knowledge_base_home (not the type-specific subdir)."""
+
+    def test_filesystem_root_is_kb_home(self, real_agent_config, mock_llm_create):
+        node = _create_node(real_agent_config)
+        expected = str(real_agent_config.path_manager.knowledge_base_home)
+
+        assert node.filesystem_func_tool is not None
+        assert node.filesystem_func_tool.config.root_path == expected
+        assert node.filesystem_func_tool._path_normalizer is not None
+
+        ns = real_agent_config.current_namespace
+        assert node.filesystem_func_tool._path_normalizer("q_001.yaml", None) == f"sql_summaries/{ns}/q_001.yaml"
